iptv_manager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
channels==4.0.0
daphne==4.0.0
django-cors-headers==4.3.0
psutil==5.9.6
python-dotenv==1.0.0
